requirements.txt
gradio==4.44.1
pydantic==2.9.2
pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
opencv-python-headless==4.10.0.84
transformers==4.44.2
torch==2.4.1
langgraph==0.2.35